:recycle: refactor: Move RedisLock into LockService
