+++ src/main/kotlin/com/team6/ticketbook/domain/book/service/BookService.kt
@@ -12,7 +12,6 @@ import com.team6.ticketbook.domain.lock.service.LockService
 import com.team6.ticketbook.domain.seat.repository.SeatRepository
 import com.team6.ticketbook.domain.show.model.Show
 import com.team6.ticketbook.domain.show.repository.ShowRepository
-import com.team6.ticketbook.infra.redis.RedisLock
 import org.springframework.data.repository.findByIdOrNull
 import org.springframework.stereotype.Service
 import org.springframework.transaction.annotation.Transactional
@@ -22,7 +21,6 @@ class BookService(
     private val bookRepository: BookRepository,
     private val showRepository: ShowRepository,
     private val seatRepository: SeatRepository,
-    private val redisLock: RedisLock,
     private val lockService: LockService,
 ) {
     fun getBookById(memberId: Long, bookId: Long): BookResponse {
@@ -55,11 +53,20 @@ class BookService(
 
     @Transactional
     fun createBookWithRedisLock(memberId: Long, request: CreateBookRequest): BookResponse =
-        "lock:${request.showId}-${request.date}-${request.seatId}".let { key ->
-            redisLock.runExclusive(key) {
+        createLockKeyWithBookRequest(request).let { key ->
+            lockService.runExclusive(key) {
                 createBook(memberId, request)
             }
         }
+
+    @Transactional
+    fun createBookWithJpaLock(memberId: Long, request: CreateBookRequest): BookResponse {
+        createLockKeyWithBookRequest(request).let {
+            lockService.jpaLock(it)
+        }
+        val bookResponse = createBook(memberId, request)
+        return bookResponse
+    }
 
     @Transactional
     fun deleteBookById(memberId: Long, bookId: Long) {
@@ -78,12 +85,8 @@ class BookService(
         }
     }
 
-    @Transactional
-    fun createBookWithMySqlLock(memberId: Long, request: CreateBookRequest): BookResponse {
-        "${request.showId}-${request.seatId}-${request.date}".let {
-            lockService.lock(it)
-        }
-        val bookResponse = createBook(memberId, request)
-        return bookResponse
+    private fun createLockKeyWithBookRequest(request: CreateBookRequest): String {
+        return "${request.showId}-${request.date}-${request.seatId}"
     }
+
 }
